# src/main/java/com/gmail/speedwagon/dto/user/BaseUserResponse.java
package com.gmail.speedwagon.dto.user;

import com.gmail.speedwagon.enums.Role;
import lombok.Getter;
import lombok.Setter;

import java.util.Set;

@Getter
@Setter
public class BaseUserResponse {
    private Long id;
    private String email;
    private String firstName;
    private Set<Role> roles;
    private String provider;
}